moved Cohort Modal to components section
