+++ src/components/CohortModal/components/CohortDetails/components/index.js
