+++ src/components/CohortModal/components/CohortDetails/index.js
