+++ src/components/CohortModal/components/shared/AlertManager.js
