+++ src/components/CohortModal/components/shared/DeleteConfirmationModal.js
