+++ src/components/CohortModal/components/shared/index.js
