+++ src/components/CohortModal/config.js
