+++ src/components/CohortModal/styles.js
